Copy frontend from old project, connect backend to frontend (#12)
* Replace Biome with Prettier & ESLint, start adapting controllers and Pinia stores

* Improve README.md

* Refactor and improve session management and other services

* Start migrating auth, country API and types

* Get the website working; more features and fixes

* Implement SQL e-mail submission

* More migrations and improvements; translate the homepage

* Fix the scope of translations

* Make the navigation bar of glass texture

* Remove extra .gitignore from frontend directory, update Dockerfiles

* Improve landing page theme

* Improve the landing page's quality even more
- reduce navbar content size
- add a background gradient

* Fix landing page consistency between light and dark theme

* Adapt the login page, improve and refactor UI

* Fix global locale switch; upgrade login and signup UI, improve misc UI

* Start defining admin account and blog posts

* Make the signup view more consistent with the login

* Replace gray with slate, fix input borders in Login and Signup

* Miscellaneous public UI improvements

* Start implementing blog posts; continue fixing compilation errors and migrating

* Migrate more pages, enable more features
- change user first name, last name, theme, language
- partially migrate User Settings and Dashboard
- make TailwindCSS styles more consistent
- add redis rate limiter with setup of 50 requests per 4 minutes
- add helmet.js for security

* Add more or update tables and triggers for constraints, divide the large SQL script

* Add GitHub workflow for main branch pushes

* Drop mentions of Firebase from the codebase

* Add dev and prod Docker compose profiles, fix auth tokens,

* Improve dashboard and migrate organization creation

* Improve the landing page and darken the background

* Implement blog posts

* Add warning note in README.md

+++ backend/src/routes/organization/member/controller.ts
@@ -1,23 +1,21 @@
-import type { Request, Response } from "express";
-import { matchedData } from "express-validator";
-import { StatusCodes } from "http-status-codes";
-import type {
-  ErrorResponse,
-  OrganizationLocals,
-} from "../../../types/data-transfer-objects.js";
-import database from "../../../services/database.js";
-import type { OrganizationMember } from "../../../types/database-types.js";
-import { camelCaseify } from "../../../utils/database.js";
-import redis from "../../../services/redis.js";
-import { REDIS_ORGANIZATION_MEMBERS } from "../../../utils/constants.js";
+import type { Request, Response } from 'express';
+import { matchedData } from 'express-validator';
+import { StatusCodes } from 'http-status-codes';
+import type { ErrorResponse, OrganizationLocals } from '../../../types/data-transfer-objects.js';
+import type { OrganizationMember } from '../../../types/database-types.js';
+import { camelCaseify } from '../../../utils/database.js';
+import { REDIS_ORGANIZATION_MEMBERS } from '../../../utils/constants.js';
+import getRedis from '../../../services/redis.js';
+import getDatabase from '../../../services/database.js';
 
+// Deprecated
 export async function GET(
   _request: Request,
-  response: Response<OrganizationMember[], OrganizationLocals>
+  response: Response<OrganizationMember[], OrganizationLocals>,
 ): Promise<any> {
-  const { rows } = await database.query(
-    "SELECT * FROM organization_member WHERE organization_id = $1",
-    [response.locals.organization.organizationId]
+  const { rows } = await getDatabase().query(
+    'SELECT * FROM organization_member WHERE organization_id = $1',
+    [response.locals.organization.organizationId],
   );
 
   response.send(rows.map<OrganizationMember>(camelCaseify));
@@ -25,7 +23,7 @@ export async function GET(
 
 export async function PATCH(
   request: Request,
-  response: Response<ErrorResponse, OrganizationLocals>
+  response: Response<ErrorResponse, OrganizationLocals>,
 ): Promise<any> {
   const { roleIds, userId } = matchedData<{
     roleIds: number[];
@@ -33,11 +31,11 @@ export async function PATCH(
   }>(request);
   const organizationId = response.locals.organization.organizationId;
 
-  await database.query(
+  await getDatabase().query(
     `INSERT INTO organization_member_role (user_id, organization_id, role_id)
     SELECT $1, $2, UNNEST($3::bigint[])
     ON CONFLICT DO NOTHING;`,
-    [userId, organizationId, roleIds]
+    [userId, organizationId, roleIds],
   );
 
   response.sendStatus(StatusCodes.OK);
@@ -46,19 +44,16 @@ export async function PATCH(
 // TODO: the member with permission or the member themselves can only invoke this
 export async function DELETE(
   request: Request,
-  response: Response<ErrorResponse, OrganizationLocals>
+  response: Response<ErrorResponse, OrganizationLocals>,
 ): Promise<any> {
   const { userId, organizationId } = matchedData<{
     userId: number;
     organizationId: number;
   }>(request);
-  const { rowCount } = await database.query(
-    "DELETE FROM member WHERE user_id = $1",
-    [userId]
-  );
+  const { rowCount } = await getDatabase().query('DELETE FROM member WHERE user_id = $1', [userId]);
 
   if (rowCount) {
-    redis.sRem(REDIS_ORGANIZATION_MEMBERS(organizationId), userId.toString());
+    getRedis().sRem(REDIS_ORGANIZATION_MEMBERS(organizationId), userId.toString());
     return response.sendStatus(StatusCodes.OK);
   }
